Change size() to return size_t

+++ org/antlr/v4/runtime/ANTLRInputStream.cpp
@@ -155,7 +155,7 @@ namespace org {
                     return p;
                 }
 
-                int ANTLRInputStream::size() {
+                size_t ANTLRInputStream::size() {
                     return n;
                 }
